% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions.R
\name{getQIIME}
\alias{getQIIME}
\title{Import from QIIME2 function}
\usage{
getQIIME(features, metadata, ASV, taxref)
}
\arguments{
\item{features}{= (required) file output from QIIME2 implementation of DADA2, as in "table.qza"}

\item{metadata}{= (required) text file (and path to file) with metadata to add to phyloseq object. Column with sample identifiers must be labeled "Sample_ID";}

\item{ASV}{= (required) QIIME2 file with dada2 ASV output. As in: "representative_sequences.qza"}

\item{taxref}{= (required) path to the training file for taxonomy assignment. We recommend using SILVA.}
}
\description{
A function that takes output from QIIME2 pipeline and outputs a phyloseq object. Assumes that DADA2 has been run in QIIME2, but you will assign taxonomy in R.
}
\examples{
QSeq(ps, abundance="QPCR_16S")
getQIIME()
}
\keyword{QIIME2,}
\keyword{phyloseq}
