% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions.R
\name{Qscale}
\alias{Qscale}
\title{Function that scales otu table by column in metadata}
\usage{
Qscale(ps, scale)
}
\arguments{
\item{ps}{= phyloseq object with community data, and sample data with at least one column for scaling}

\item{scale}{= vector that is used for scaling samples}
}
\description{
Takes a phyloseq object, and vector for scaling samples in otu table. After scaling, rounds taxon count values to the nearest integer, then returns a phyloseq object with scaled community data.
}
\examples{
QScale()
}
\keyword{QSeq,}
\keyword{Quantitative}
\keyword{Scaling}
\keyword{Sequencing,}
