% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions.R
\name{QSeq}
\alias{QSeq}
\title{QSeq function}
\usage{
QSeq(ps, abundance)
}
\arguments{
\item{ps}{= phyloseq object with community data, and sample data with at least one column for scaling}

\item{abundance}{= name of column in metadata to scale counts by}
}
\description{
A wrapper for QScale that takes phyloseq object as input. Metadata needs to have sample column with scaling factor
}
\examples{
QSeq(ps, abundance="QPCR_16S")
QSeq()
}
\keyword{QSeq,}
\keyword{Quantitative}
\keyword{Scaling}
\keyword{Sequencing,}
